Got the App to start running and communicating with the hsqldb :D yeah!
KNOWN BUGS:
the name and email from EmailOnlyAddress and PostalAddress will not be
written to the database.

Pressing the "READ ALL"-Button will load, but not display the Data. The
Data will just be displayed, if you press the "SAVE ALL"-Button after
the "READ-ALL"-Button.  

The delete functionality still doesn't work.